change quantity type to double

+++ src/Conclave.Common/Models/Entities/ConclaveOwnerSnapshot.cs
@@ -7,7 +7,7 @@ public class ConclaveOwnerSnapshot
     public Guid Id { get; set; } = Guid.NewGuid();
     public ConclaveEpoch ConclaveEpoch { get; set; } = new();
     public DelegatorSnapshot DelegatorSnapshot { get; set; } = new();
-    public ulong Quantity { get; set; }
+    public double Quantity { get; set; }
     public DateTime DateCreated { get; set; } = DateUtils.AddOffsetToUtc(DateTime.UtcNow);
     public DateTime DateUpdated { get; set; } = DateUtils.AddOffsetToUtc(DateTime.UtcNow);
 }
